Fake for link account manager
