+++ link/src/main/java/com/stripe/android/link/injection/LinkModule.kt
@@ -5,6 +5,8 @@ import com.stripe.android.core.Logger
 import com.stripe.android.core.injection.IOContext
 import com.stripe.android.core.networking.DefaultStripeNetworkClient
 import com.stripe.android.core.version.StripeSdkVersion
+import com.stripe.android.link.account.LinkAccountManager
+import com.stripe.android.link.account.LinkAccountManagerImpl
 import com.stripe.android.link.analytics.DefaultLinkEventsReporter
 import com.stripe.android.link.analytics.LinkEventsReporter
 import com.stripe.android.link.repositories.LinkApiRepository
@@ -25,6 +27,10 @@ internal interface LinkModule {
     @Binds
     @LinkScope
     fun bindLinkEventsReporter(linkEventsReporter: DefaultLinkEventsReporter): LinkEventsReporter
+
+    @Binds
+    @LinkScope
+    fun bindLinkAccountManager(linkAccountManager: LinkAccountManagerImpl): LinkAccountManager
 
     companion object {
         @Provides
